% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scope.R
\name{gen_token}
\alias{gen_token}
\title{generate oauth token}
\usage{
gen_token(app.id = Sys.getenv("WAKATIME_ID"),
  app.secret = Sys.getenv("WAKATIME_SECRET"))
}
\arguments{
\item{app.id}{character. OAuth app ID. If you set as system environment
(recommend) variables is used (check \code{Sys.setenv('WAKATIME_ID')}).}

\item{app.secret}{character. OAuth secret app key. If you set as system environment}
}
\description{
generate oauth token
}

