% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recent_files}
\alias{recent_files}
\title{Get recently opened file name}
\usage{
recent_files(n = 15)
}
\arguments{
\item{n}{integer. retern file numbers. default is 15 .(max)}
}
\value{
data frame. recent open files in RStudio session. Newer to older
}
\description{
Get recently opened file name
}

