% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakatimeR.R
\name{set_wakatime_api_key}
\alias{set_wakatime_api_key}
\title{Set WakaTime API Key}
\usage{
set_wakatime_api_key(api_key, config_path)
}
\arguments{
\item{api_key}{Character. WakaTime API key.}

\item{config_path}{Character. wakatimeR configuration file path.}
}
\description{
Set the WakaTime API key to the wakatimeR configuration file
}
