% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakatimeR.R
\name{start_wakatime}
\alias{start_wakatime}
\title{Start wakatimeR for RStudio}
\usage{
start_wakatime(
  heartbeat_interval = 120,
  polling_interval = 120,
  binary_path = NULL,
  config_path = "~/.wakatime/wakatimeR.cfg",
  log_path = "~/.wakatime/wakatimeR.log",
  log_level = logger::INFO
)
}
\arguments{
\item{heartbeat_interval}{Numeric. The interval, in seconds, for sending heartbeats to WakaTime.}

\item{polling_interval}{Numeric. The interval, in seconds, for checking file changes.}

\item{binary_path}{Character. \code{wakatime-cli} binary file path. Default: \code{NULL} which will look for the binary in the system PATH.}

\item{config_path}{Character. wakatimeR configuration file path. Default: \code{"~/.wakatime/wakatimeR.cfg"}.}

\item{log_path}{Character. Log file path. Default: \code{"~/.wakatime/wakatimeR.log"}.}

\item{log_level}{Logger level. The logging level to use. Default is \code{logger::INFO}.}
}
\description{
Starts WakaTimeR for RStudio by setting up the necessary configurations,
checking for the \code{wakatime-cli} binary, and setting up event polling.
It also ensures that the WakaTime API key is properly configured.
}
