% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{write_scope}
\alias{write_scope}
\title{Authorize using OAuth 2.0 client token.}
\usage{
write_scope(app.id = Sys.getenv("WAKATIME_ID"),
  app.secret = Sys.getenv("WAKATIME_SECRET"))
}
\arguments{
\item{app.id}{character. OAuth app ID. If you set as system environment
(recommend) variables is used (check \code{Sys.setenv('WAKATIME_ID')}).}

\item{app.secret}{character. OAuth secret app key. If you set as system environment}
}
\description{
If you not register application,
 create app and configure id and secret key from \href{https://wakatime.com/apps}{WakaTime}.
}

