% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_api.R
\name{wt_api}
\alias{wt_api}
\title{wakatime api wrapper}
\usage{
wt_api(resource = c("heartbeats", "durations", "stats", "users", "summaries",
  "user_agents"), key = Sys.getenv("WAKATIME_KEY"), param = list(date =
  format(Sys.Date(), "\%m/\%d/\%Y"), time = "time", "entity"))
}
\arguments{
\item{resource}{select request api resource type}

\item{key}{character. your wakatime api key}

\item{param}{request url parameters provide as list.}
}
\description{
wakatime api wrapper
}

